REF: Move unittest.TestResult method replacements
